requirements.txt
python-telegram-bot==20.7
google-generativeai>=1.0.0
python-dotenv==1.0.0
requests==2.31.0
google-cloud-speech==2.21.0